Facial-recognition/requirements.txt
THE LIST OF REQUIREMENTS:
h5py==2.8.0
Keras==2.2.4
tensorflow==1.15.2
dlib==19.16.0
opencv_python==3.4.3.18
imutils==0.5.1
numpy==1.15.2
matplotlib==3.0.0
scipy==1.1.0
protobuf==3.6.0
Pillow
